Move zkasm generation into zkasm_codegen.rs
